sysd: Split service into multiple submodules
